requirements.txt
requests
beautifulsoup4
lxml
csv
selenium
pillow
pytesseract
